Split app.js out of index.js for supertest
