+++ server/index.js
@@ -1,17 +1,5 @@
-require('dotenv').config();
-const express = require('express');
-const path = require('path');
+const app = require('./app.js');
 
-const routes = require('./routes');
-
-const app = express();
 const port = process.env.PORT || 3003;
-
-app.use(express.static(path.join(__dirname, '/../public')));
-
-app.use(express.json());
-app.use(express.urlencoded({ extended: true }));
-
-app.use('/api', routes);
 
 app.listen(port);
